Implement TPM Bound Key attestation
The previous implementation passed around a signed SSH Public key.
However I realized Certify_Creation makes a signature with the AIK and
we can derive the SSH key from the tpm2.TPMTPublic of the TPM Bound key.

Then we actually validate it was created on the TPM and bound by it.

Signed-off-by: Morten Linderud <morten@linderud.pw>